projects/agentic-automation-framework-python/requirements.txt
playwright==1.56.0
python-dotenv==1.1.1
requests==2.32.3
pytest==8.3.5
